QAA-5216: Switch validate PR to node shell
When validating the PR title, GitHub actions does not seem to
provide a way to escape any quotes for shell usage, so titles
with quotes just fail.
However, the event payload is also available as a JSON file
on the filesystem that can be loaded and parsed. While this
could be done in any shell, probably Node is the most familiar
language for us.